Move coredump to a separate file, use efficient read
